requirements.txt
streamlit==1.39.0
langchain==0.1.20
langchain-community==0.0.38
langchain-experimental==0.0.58
pandas==2.2.2
faiss-cpu==1.8.0
sentence-transformers==3.0.1
ollama==0.3.3
numpy==1.26.4
python-dotenv==1.0.1
sqlalchemy==2.0.31
tabulate==0.9.0
plotly==5.22.0